updated multiplication table program

+++ control-flow/multiplication_table.py
@@ -1,5 +1,5 @@
 # Multiplication table
-number = int(input("Enter a number to see its multiplicaton table: "))
+number = float(input("Enter a number to see its multiplicaton table: "))
 
 for i in range(1, 11):
     print(f"{number} * {i} = {number * i}")
